backend/requirements.txt
django>=5.1,<5.3
djangorestframework>=3.15,<3.17
drf-spectacular>=0.28,<0.29
PyMySQL>=1.1,<1.2
